requirements.txt
mysql-connector-python==8.0.22
pandas==1.1.5
requests==2.25.0
streamlit==0.74.1
